refactor: 拆分隐藏文件夹功能到独立模块
- 新增 services/hide-folders.js，从 image-paste-handler.js 提取隐藏文件夹逻辑
- 更新 input.js 的 require 引用路径
- 控制单文件行数在 150 行以内
